+++ .storybook/package.json
@@ -1,0 +1,2 @@
+{
+}
