+++ src/components/checkbox/Checkbox.stories.js
@@ -1,4 +1,4 @@
-import React from '@storybook/react'
+// import React from '@storybook/react'
 import { action } from '@storybook/addon-actions'
 
 import Checkbox from './Checkbox.js'
