+++ src/components/language-selector/LanguageSelector.stories.js
@@ -1,5 +1,5 @@
 // @ts-check
-import React from '@storybook/react'
+// import React from '@storybook/react'
 
 import i18n from '../../i18n.js'
 import LanguageSelectorEl from './LanguageSelector.js'
