+++ src/files/modals/new-folder-modal/NewFolderModal.stories.js
@@ -1,4 +1,4 @@
-import React from '@storybook/react'
+// import React from '@storybook/react'
 import { action } from '@storybook/addon-actions'
 import i18n from '../../../i18n-decorator.js'
 import NewFolderModal from './NewFolderModal.js'
