+++ src/files/selected-actions/SelectedActions.stories.js
@@ -1,4 +1,4 @@
-import React from '@storybook/react'
+// import React from '@storybook/react'
 import { withKnobs, number } from '@storybook/addon-knobs'
 import { action } from '@storybook/addon-actions'
 import i18n from '../../i18n-decorator.js'
