+++ src/loader/ComponentLoader.stories.js
@@ -1,4 +1,4 @@
-import React from '@storybook/react'
+// import React from '@storybook/react'
 import ComponentLoader from './ComponentLoader.js'
 
 /**
